+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## [Unreleased]
 ### Added
+- Added mechanism to refresh list of schematics. Should be a button next to the search bar on panel showing list of schematics.
 
 ### Changed